+++ CHANGELOG.md
@@ -34,3 +34,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Exists Script
 - Fetch publications and other documents
 - Documents Table
+- project-code as parameter in project-view